Feature/staging timeout timing grafana

+++ deploy/prod/prod-stop.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
-echo "Stop running docker compose"
-cd A2rchi-prod/deploy/prod/
-docker compose -f prod-compose.yaml down
+if [ -z "$(ls -A A2rchi-prod/deploy/prod/)" ]; then
+    echo "Deployment directory is empty; skipping docker compose down"
+else
+    echo "Stop running docker compose"
+    cd A2rchi-prod/deploy/prod/
+    docker compose -f prod-compose.yaml down
+fi
